pyproject.toml
[project]
name = "vision-llm"
version = "0.2.0"
description = "Vector path-based visual language model with diffusion"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "fonttools>=4.0",
    "Pillow>=10.0",
    "pytesseract>=0.3",
    "numpy>=1.24,<2",
]

[project.optional-dependencies]
dev = ["pytest", "matplotlib"]
